refactor(app): move nbt generation into worker
